reuse push command with --manifest flag instead of new upload command
