--- cli/commands/upload/index.ts
@@ -1,30 +1,0 @@
-import { CommandHandler } from '../..'
-import { assertExists } from '../../utils'
-import { AppendToFile } from '../../utils/append.to.file'
-import { UploadManifest } from '../publish/upload.manifest'
-import { GetCredentials } from '../../utils/get.credentials'
-import { UploadImage } from '../publish/upload.image'
-
-export default function provideUpload(
-  getCredentials: GetCredentials,
-  uploadImage: UploadImage,
-  uploadManifest: UploadManifest,
-  appendToFile: AppendToFile,
-  args: any
-): CommandHandler {
-  assertExists(getCredentials, 'getCredentials')
-  assertExists(uploadImage, 'uploadImage')
-  assertExists(uploadManifest, 'uploadManifest')
-  assertExists(appendToFile, 'appendToFile')
-  assertExists(args, 'args')
-
-  return async function upload() {
-    const imageReference = await uploadImage()
-    const { privateKey } = await getCredentials()
-    const manifestReference = await uploadManifest(imageReference, privateKey)
-
-    let logMessage = `${new Date().toUTCString()}: successfully uploaded manifest with reference ${manifestReference}`
-    console.log(logMessage)
-    appendToFile(logMessage, 'publish.log')
-  } 
-}
